move almost all types to types.rs
